refactor: édition des commentaires et du code de l'update des todo
